requirements.txt
flask
flask-Bootstrap4
pyserial
flask-wtf
firebase-admin